--- src/main/resources/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-CN">
-<head>
-    <title>Webpack App</title>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width; initial-scale=1.0; maximum-scale=1.0; user-scalable=0;">
-    <link rel="stylesheet" href="/dist/main.css">
-</head>
-<body>
-    <div id="app"></div>
-    <script type="text/javascript" src="/dist/vendors.js"></script>
-    <script type="text/javascript" src="/dist/main.js"></script>
-</body>
-</html>
