--- src/main/resources/iview/config/config.js
@@ -1,9 +1,0 @@
-/**
- * Created by aresn on 16/7/18.
- */
-import Env from './env';
-
-let config = {
-    env: Env
-};
-export default config;
