--- src/main/resources/iview/config/env.js
@@ -1,1 +1,0 @@
-export default "development";
