--- src/main/resources/iview/libs/util.js
@@ -1,11 +1,0 @@
-/**
- * Created by aresn on 16/7/18.
- */
-let util = {
-
-};
-util.alert = function(content) {
-    window.alert(content);
-};
-
-export default util;
