--- src/main/resources/iview/views/index.vue
@@ -1,14 +1,0 @@
-<template>
-    <Row type="flex" justify="center" align="middle">
-        <i-col span="24">
-            <h1>
-                <img src="../images/logo.png">
-            </h1>
-        </i-col>
-    </Row>
-</template>
-<script>
-    export default {
-
-    }
-</script>
