--- src/main/resources/package.json
@@ -1,56 +1,0 @@
-{
-  "name": "datarest-project",
-  "version": "1.0.2",
-  "description": "A base project using Vue.js1.0、Vue-Router、webpack and with iView.",
-  "main": "index.js",
-  "scripts": {
-    "init": "webpack --progress --config webpack.dev.config.js",
-    "dev": "webpack-dev-server --inline --hot --compress --history-api-fallback --config webpack.dev.config.js",
-    "build": "webpack --progress --hide-modules --config webpack.prod.config.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/iview/iview-project.git"
-  },
-  "author": "Kevin",
-  "license": "MIT",
-  "dependencies": {
-    "vue": "^1.0.26",
-    "vue-resource": "^1.2.0",
-    "vue-router": "^0.7.13",
-    "vue-validator": "^3.0.0-alpha.2",
-    "iview": "^1.0.1"
-  },
-  "devDependencies": {
-    "autoprefixer-loader": "^2.0.0",
-    "babel": "^6.3.13",
-    "babel-core": "^6.3.21",
-    "babel-loader": "^6.2.0",
-    "babel-plugin-transform-runtime": "^6.3.13",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-runtime": "^5.8.34",
-    "css-loader": "^0.23.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
-    "html-loader": "^0.3.0",
-    "html-webpack-plugin": "^2.22.0",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "node-sass": "^3.10.1",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "vue-hot-reload-api": "^1.2.0",
-    "vue-html-loader": "^1.2.0",
-    "vue-loader": "^7.2.0",
-    "vue-style-loader": "^1.0.0",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.16.1",
-    "underscore": "^1.8.3",
-    "axios": "^0.15.3"
-  },
-  "bugs": {
-    "url": "https://github.com/iview/iview-project/issues"
-  },
-  "homepage": "https://www.iviewui.com"
-}
