+++ src/main/webapp/.babelrc
@@ -1,0 +1,5 @@
+{
+  "presets": [
+    ["es2015", { "modules": false }]
+  ]
+}
