+++ src/main/webapp/src/views/queryStatement/stepPreview.vue
@@ -1,0 +1,32 @@
+<template>
+    <Modal v-model="localShowModal" title="SQL模板预览" :closable="false">
+        <div style="text-align:center">
+            <p> SELECT {{ saveModel.selectBody }}
+                FROM {{ saveModel.fromBody }}
+                WHERE {{ saveModel.whereBody }}
+            </p>
+        </div>
+        <p slot="footer">
+            <Button type="ghost" @click="cancel">返回</Button>&nbsp;&nbsp;
+        </p>
+    </Modal>
+</template>
+<script>
+    export default {
+        data () {
+            return {
+                localShowModal: false,
+                saveModel: {}
+            }
+        },
+        methods: {
+            show (showModal, saveModel) {
+                this.localShowModal = showModal
+                this.saveModel = saveModel
+            },
+            cancel () {
+                this.localShowModal = false
+            }
+        }
+    }
+</script>
